moved get, store method to context
